mathx_robotics/pyproject.toml
[build-system]
requires = [
  "setuptools"
]
build-backend = "setuptools.build_meta"
install_requires=[]

[project]
name="mathx-robotics"
version="1.0.0"
requires-python = "<3.12,>=3.11" #dm-reverb only works with older python
dependencies=[
  "numpy", #Tensorflow version below needs an older version of numpy
  "mujoco",
  #"jax==0.4.34",
  #"jaxlib==0.5.1",
  #"jax[cuda]==0.5.1",
  #"jax==0.5.1",
  "jax",
  #"jax[cuda12]==0.4.34", #For some reason this version is much faster.
  "optax",
  "flax",
  #"opencv-python",
  "gymnasium[classic_control]",
  "matplotlib",
  "pyqt6", #Needed for interactive plots
  "natsort",
  "fiblat<0.6",
  "rlax", #Needed to avoid dm-acme/dm-env installing rlax from the repo
  "dm-haiku",
  #"dm-haiku @ git+https://github.com/google-deepmind/dm-haiku.git",
  #"dm-reverb[tensorflow]", #This forces a compatible version of tensorflow to be installed
  #"dm-reverb @ git+https://github.com/google-deepmind/reverb.git",
  #"tf-nightly",
  #"dm-reverb-nightly[tensorflow]",
  "dm-env",
  #"dm-env @ git+https://github.com/google-deepmind/dm_env.git",
  "dm-control",
  #"dm-control @ git+https://github.com/google-deepmind/dm_control.git",
  "dm-acme",
  #"dm-acme @ git+https://github.com/google-deepmind/acme.git",
  "launchpad",
  #"gymnasium",
  "stable-baselines3",
  "imageio"
]
